Yesod.Content

Content

data Content

Constructors

ContentBuilder !Builder !(Maybe Int)
The content and optional content length.

ContentSource !(Source (ResourceT IO) (Flush Builder))

ContentFile !FilePath !(Maybe FilePart)

ContentDontEvaluate !Content

emptyContent :: Content

Zero-length enumerator.

class ToContent a where

Anything which can be converted into Content. Most of the time, you will want to use the ContentBuilder constructor. An easier approach will be to use a pre-defined toContent function, such as converting your data into a lazy bytestring and then calling toContent on that.

Please note that the built-in instances for lazy data structures (String, lazy ByteString, lazy Text and Html) will not automatically include the content length for the ContentBuilder constructor.

Methods

toContent :: a -> Content

Mime types

Data type

type ContentType = ByteString

typeHtml :: ContentType

typePlain :: ContentType

typeJson :: ContentType

typeXml :: ContentType

typeAtom :: ContentType

typeRss :: ContentType

typeJpeg :: ContentType

typePng :: ContentType

typeGif :: ContentType

typeSvg :: ContentType

typeJavascript :: ContentType

typeCss :: ContentType

typeFlv :: ContentType

typeOgv :: ContentType

typeOctet :: ContentType

Utilities

simpleContentType :: ContentType -> ContentType

Removes "extra" information at the end of a content type string. In particular, removes everything after the semicolon, if present.

For example, "text/html; charset=utf-8" is commonly used to specify the character encoding for HTML data. This function would return "text/html".

Evaluation strategy

newtype DontFullyEvaluate a

Prevents a response body from being fully evaluated before sending the request.

Since 1.1.0

Constructors

DontFullyEvaluate a

Representations

type ChooseRep

Arguments

= [ContentType]
list of content-types user accepts, ordered by preference

-> IO (ContentType, Content)

A function which gives targetted representations of content based on the content-types the user accepts.

class HasReps a where

Any type which can be converted to representations.

Methods

chooseRep :: a -> ChooseRep

defChooseRep :: [(ContentType, a -> IO Content)] -> a -> ChooseRep

A helper method for generating HasReps instances.

This function should be given a list of pairs of content type and conversion functions. If none of the content types match, the first pair is used.

Specific content types

newtype RepHtml

Constructors

RepHtml Content

newtype RepJson

Constructors

RepJson Content

data RepHtmlJson

Constructors

RepHtmlJson Content Content

newtype RepPlain

Constructors

RepPlain Content

newtype RepXml

Constructors

RepXml Content

Utilities

formatW3 :: UTCTime -> Text

Format a UTCTime in W3 format.

formatRFC1123 :: UTCTime -> Text

Format as per RFC 1123.

formatRFC822 :: UTCTime -> Text

Format as per RFC 822.