Yesod.Core.Content

Content

data Content

Constructors

ContentBuilder !Builder !(Maybe Int)
The content and optional content length.

ContentSource !(ConduitT () (Flush Builder) (ResourceT IO) ())

ContentFile !FilePath !(Maybe FilePart)

ContentDontEvaluate !Content

emptyContent :: Content

Zero-length enumerator.

class ToContent a where

Anything which can be converted into Content. Most of the time, you will want to use the ContentBuilder constructor. An easier approach will be to use a pre-defined toContent function, such as converting your data into a lazy bytestring and then calling toContent on that.

Please note that the built-in instances for lazy data structures (String, lazy ByteString, lazy Text and Html) will not automatically include the content length for the ContentBuilder constructor.

Methods

toContent :: a -> Content

class ToFlushBuilder a where

A class for all data which can be sent in a streaming response. Note that for textual data, instances must use UTF-8 encoding.

Since 1.2.0

Methods

toFlushBuilder :: a -> Flush Builder

Mime types

Data type

type ContentType = ByteString

typeHtml :: ContentType

typePlain :: ContentType

typeJson :: ContentType

typeXml :: ContentType

typeAtom :: ContentType

typeRss :: ContentType

typeJpeg :: ContentType

typePng :: ContentType

typeGif :: ContentType

typeSvg :: ContentType

typeJavascript :: ContentType

typeCss :: ContentType

typeFlv :: ContentType

typeOgv :: ContentType

typeOctet :: ContentType

Utilities

simpleContentType :: ContentType -> ContentType

Removes "extra" information at the end of a content type string. In particular, removes everything after the semicolon, if present.

For example, "text/html; charset=utf-8" is commonly used to specify the character encoding for HTML data. This function would return "text/html".

contentTypeTypes :: ContentType -> (ByteString, ByteString)

Give just the media types as a pair.

For example, "text/html; charset=utf-8" returns ("text", "html")

Evaluation strategy

newtype DontFullyEvaluate a

Prevents a response body from being fully evaluated before sending the request.

Since 1.1.0

Constructors

DontFullyEvaluate

Fields

unDontFullyEvaluate :: a

Representations

data TypedContent

Constructors

TypedContent !ContentType !Content

class ToContent a => ToTypedContent a where

Any type which can be converted to TypedContent.

Since 1.2.0

Methods

toTypedContent :: a -> TypedContent

class ToTypedContent a => HasContentType a where

Methods

getContentType :: Monad m => m a -> ContentType

Specific content types

type RepHtml = Html

Deprecated: Please use Html instead

newtype RepJson

Constructors

RepJson Content

newtype RepPlain

Constructors

RepPlain Content

newtype RepXml

Constructors

RepXml Content

Smart constructors

repJson :: ToContent a => a -> RepJson

repPlain :: ToContent a => a -> RepPlain

repXml :: ToContent a => a -> RepXml