Yesod.Core.Json

Convert from a JSON value

defaultLayoutJson

Arguments

:: (Yesod site, ToJSON a)

=> WidgetFor site ()
HTML

-> HandlerFor site a
JSON

-> HandlerFor site TypedContent

Provide both an HTML and JSON representation for a piece of data, using the default layout for the HTML output (defaultLayout).

Since: 0.3.0

jsonToRepJson :: (Monad m, ToJSON a) => a -> m Value

Deprecated: Use returnJson instead

Wraps a data type in a RepJson. The data type must support conversion to JSON via ToJSON.

Since: 0.3.0

returnJson :: (Monad m, ToJSON a) => a -> m Value

Convert a value to a JSON representation via aeson's toJSON function.

Since: 1.2.1

returnJsonEncoding :: (Monad m, ToJSON a) => a -> m Encoding

Convert a value to a JSON representation via aeson's toEncoding function.

Since: 1.4.21

provideJson :: (Monad m, ToJSON a) => a -> Writer (Endo [ProvidedRep m]) ()

Provide a JSON representation for usage with selectReps, using aeson's toJSON (aeson >= 0.11: toEncoding) function to perform the conversion.

Since: 1.2.1

Convert to a JSON value

parseCheckJsonBody :: (MonadHandler m, FromJSON a) => m (Result a)

Parse the request body to a data type as a JSON value. The data type must support conversion from JSON via FromJSON. If you want the raw JSON value, just ask for a Result Value.

The MIME type must indicate JSON content. Requiring a JSON content-type helps secure your site against CSRF attacks (browsers will perform POST requests for form and text/plain content-types without doing a CORS check, and those content-types can easily contain valid JSON).

Note that this function will consume the request body. As such, calling it twice will result in a parse error on the second call, since the request body will no longer be available.

Since: 0.3.0

parseInsecureJsonBody :: (MonadHandler m, FromJSON a) => m (Result a)

Same as parseCheckJsonBody, but does not check that the mime type indicates JSON content.

Note: This function is vulnerable to CSRF attacks.

Since: 1.6.11

requireCheckJsonBody :: (MonadHandler m, FromJSON a) => m a

Same as parseCheckJsonBody, but return an invalid args response on a parse error.

requireInsecureJsonBody :: (MonadHandler m, FromJSON a) => m a

Same as parseInsecureJsonBody, but return an invalid args response on a parse error.

Since: 1.6.11

Deprecated JSON conversion

parseJsonBody :: (MonadHandler m, FromJSON a) => m (Result a)

Deprecated: Use parseCheckJsonBody or parseInsecureJsonBody instead

Same as parseInsecureJsonBody

Since: 0.3.0

parseJsonBody_ :: (MonadHandler m, FromJSON a) => m a

Deprecated: Use requireCheckJsonBody or requireInsecureJsonBody instead

Same as parseInsecureJsonBody, but return an invalid args response on a parse error.

requireJsonBody :: (MonadHandler m, FromJSON a) => m a

Deprecated: Use requireCheckJsonBody or requireInsecureJsonBody instead

Same as parseInsecureJsonBody, but return an invalid args response on a parse error.

Produce JSON values

data Value

A JSON value represented as a Haskell value.

Constructors

Object !Object

Array !Array

String !Text

Number !Scientific

Bool !Bool

Null

class ToJSON a where

A type that can be converted to JSON.

Instances in general must specify toJSON and should (but don't need to) specify toEncoding.

An example type and instance:

-- Allow ourselves to write Text literals.
{-# LANGUAGE OverloadedStrings #-}

data Coord = Coord { x :: Double, y :: Double }

instance ToJSON Coord where
  toJSON (Coord x y) = object ["x" .= x, "y" .= y]

  toEncoding (Coord x y) = pairs ("x" .= x <> "y" .= y)

Instead of manually writing your ToJSON instance, there are two options to do it automatically:

Data.Aeson.TH provides Template Haskell functions which will derive an instance at compile time. The generated instance is optimized for your type so it will probably be more efficient than the following option.

The compiler can provide a default generic implementation for toJSON.

To use the second, simply add a deriving Generic clause to your datatype and declare a ToJSON instance. If you require nothing other than defaultOptions, it is sufficient to write (and this is the only alternative where the default toJSON implementation is sufficient):

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics

data Coord = Coord { x :: Double, y :: Double } deriving Generic

instance ToJSON Coord where
    toEncoding = genericToEncoding defaultOptions

If on the other hand you wish to customize the generic decoding, you have to implement both methods:

customOptions = defaultOptions
                { fieldLabelModifier = map toUpper
                }

instance ToJSON Coord where
    toJSON     = genericToJSON customOptions
    toEncoding = genericToEncoding customOptions

Previous versions of this library only had the toJSON method. Adding toEncoding had two reasons:

toEncoding is more efficient for the common case that the output of toJSON is directly serialized to a ByteString. Further, expressing either method in terms of the other would be non-optimal.

The choice of defaults allows a smooth transition for existing users: Existing instances that do not define toEncoding still compile and have the correct semantics. This is ensured by making the default implementation of toEncoding use toJSON. This produces correct results, but since it performs an intermediate conversion to a Value, it will be less efficient than directly emitting an Encoding. (this also means that specifying nothing more than instance ToJSON Coord would be sufficient as a generically decoding instance, but there probably exists no good reason to not specify toEncoding in new instances.)

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

Convert a Haskell value to a JSON-friendly intermediate type.

toEncoding :: a -> Encoding

Encode a Haskell value as JSON.

The default implementation of this method creates an intermediate Value using toJSON. This provides source-level compatibility for people upgrading from older versions of this library, but obviously offers no performance advantage.

To benefit from direct encoding, you must provide an implementation for this method. The easiest way to do so is by having your types implement Generic using the DeriveGeneric extension, and then have GHC generate a method body as follows.

instance ToJSON Coord where
    toEncoding = genericToEncoding defaultOptions

toJSONList :: [a] -> Value

toEncodingList :: [a] -> Encoding

class FromJSON a where

A type that can be converted from JSON, with the possibility of failure.

In many cases, you can get the compiler to generate parsing code for you (see below). To begin, let's cover writing an instance by hand.

There are various reasons a conversion could fail. For example, an Object could be missing a required key, an Array could be of the wrong size, or a value could be of an incompatible type.

The basic ways to signal a failed conversion are as follows:

fail yields a custom error message: it is the recommended way of reporting a failure;

empty (or mzero) is uninformative: use it when the error is meant to be caught by some (<|>);

typeMismatch can be used to report a failure when the encountered value is not of the expected JSON type; unexpected is an appropriate alternative when more than one type may be expected, or to keep the expected type implicit.

prependFailure (or modifyFailure) add more information to a parser's error messages.

An example type and instance using typeMismatch and prependFailure:

-- Allow ourselves to write Text literals.
{-# LANGUAGE OverloadedStrings #-}

data Coord = Coord { x :: Double, y :: Double }

instance FromJSON Coord where
    parseJSON (Object v) = Coord
        <$> v .: "x"
        <*> v .: "y"

    -- We do not expect a non-Object value here.
    -- We could use empty to fail, but typeMismatch
    -- gives a much more informative error message.
    parseJSON invalid    =
        prependFailure "parsing Coord failed, "
            (typeMismatch "Object" invalid)

For this common case of only being concerned with a single type of JSON value, the functions withObject, withScientific, etc. are provided. Their use is to be preferred when possible, since they are more terse. Using withObject, we can rewrite the above instance (assuming the same language extension and data type) as:

instance FromJSON Coord where
    parseJSON = withObject "Coord" $ \v -> Coord
        <$> v .: "x"
        <*> v .: "y"

Instead of manually writing your FromJSON instance, there are two options to do it automatically:

Data.Aeson.TH provides Template Haskell functions which will derive an instance at compile time. The generated instance is optimized for your type so it will probably be more efficient than the following option.

The compiler can provide a default generic implementation for parseJSON.

To use the second, simply add a deriving Generic clause to your datatype and declare a FromJSON instance for your datatype without giving a definition for parseJSON.

For example, the previous example can be simplified to just:

{-# LANGUAGE DeriveGeneric #-}

import GHC.Generics

data Coord = Coord { x :: Double, y :: Double } deriving Generic

instance FromJSON Coord

The default implementation will be equivalent to parseJSON = genericParseJSON defaultOptions; if you need different options, you can customize the generic decoding by defining:

customOptions = defaultOptions
                { fieldLabelModifier = map toUpper
                }

instance FromJSON Coord where
    parseJSON = genericParseJSON customOptions

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONList :: Value -> Parser [a]

array :: ToJSON a => [a] -> Value

Convert a list of values to an Value.

object :: [Pair] -> Value

Create a Value from a list of name/value Pairs. If duplicate keys arise, earlier keys and their associated values win.

(.=) :: (KeyValue kv, ToJSON v) => Text -> v -> kv infixr 8

(.:) :: FromJSON a => Object -> Text -> Parser a

Retrieve the value associated with the given key of an Object. The result is empty if the key is not present or the value cannot be converted to the desired type.

This accessor is appropriate if the key and value must be present in an object for it to be valid. If the key and value are optional, use .:? instead.

Convenience functions

jsonOrRedirect

Arguments

:: (MonadHandler m, ToJSON a)

=> Route (HandlerSite m)
Redirect target

-> a
Data to send via JSON

-> m Value

jsonOrRedirect simplifies the scenario where a POST handler sends a different response based on Accept headers:

200 with JSON data if the client prefers application/json (e.g. AJAX, see acceptsJSON).

3xx otherwise, following the PRG pattern.

jsonEncodingOrRedirect

Arguments

:: (MonadHandler m, ToJSON a)

=> Route (HandlerSite m)
Redirect target

-> a
Data to send via JSON

-> m Encoding

jsonEncodingOrRedirect simplifies the scenario where a POST handler sends a different response based on Accept headers:

200 with JSON data if the client prefers application/json (e.g. AJAX, see acceptsJSON).

3xx otherwise, following the PRG pattern. @since 1.4.21

acceptsJson :: MonadHandler m => m Bool

Returns True if the client prefers application/json as indicated by the Accept HTTP header.