Yesod.Internal.TestApi

Description

WARNING: This module exposes internal interfaces solely for the purpose of facilitating cabal-driven testing of said interfaces. This module is NOT part of the public Yesod API and should NOT be imported by library users.

Documentation

randomString :: RandomGen g => Int -> g -> String

Generate a random String of alphanumerical characters (a-z, A-Z, and 0-9) of the given length using the given random number generator.

parseWaiRequest'

Arguments

:: RandomGen g

=> Request

-> [(Text, ByteString)]
session

-> Bool

-> Word64

-> g

-> Request