Yesod.Core.Types

Documentation

type SessionMap = Map Text ByteString

type SaveSession

Arguments

= SessionMap
The session contents after running the handler

-> IO [Header]

newtype SessionBackend

Constructors

SessionBackend

Fields

sbLoadSession :: Request -> IO (SessionMap, SaveSession)

Return the session data and a function to save the session

data SessionCookie

Constructors

SessionCookie !(Either UTCTime ByteString) !ByteString !SessionMap

data ClientSessionDateCache

Constructors

ClientSessionDateCache

Fields

csdcNow :: !UTCTime

csdcExpires :: !UTCTime

csdcExpiresSerialized :: !ByteString

data YesodRequest

The parsed request information. This type augments the standard WAI Request with additional information.

Constructors

YesodRequest

Fields

reqGetParams :: ![(Text, Text)]

Same as queryString, but decoded to Text.

reqCookies :: ![(Text, Text)]

reqWaiRequest :: !Request

reqLangs :: ![Text]

Languages which the client supports. This is an ordered list by preference.

reqToken :: !(Maybe Text)

A random, session-specific token used to prevent CSRF attacks.

reqSession :: !SessionMap

Initial session sent from the client.

Since 1.2.0

reqAccept :: ![ContentType]

An ordered list of the accepted content types.

Since 1.2.0

data YesodResponse

An augmented WAI Response. This can either be a standard Response, or a higher-level data structure which Yesod will turn into a Response.

Constructors

YRWai !Response

YRWaiApp !Application

YRPlain !Status ![Header] !ContentType !Content !SessionMap

type RequestBodyContents = ([(Text, Text)], [(Text, FileInfo)])

A tuple containing both the POST parameters and submitted files.

data FileInfo

Constructors

FileInfo

Fields

fileName :: !Text

fileContentType :: !Text

fileSourceRaw :: !(ConduitT () ByteString (ResourceT IO) ())

fileMove :: !(FilePath -> IO ())

data FileUpload

Constructors

FileUploadMemory !(BackEnd ByteString)

FileUploadDisk !(InternalState -> BackEnd FilePath)

FileUploadSource !(BackEnd (ConduitT () ByteString (ResourceT IO) ()))

data Approot master

How to determine the root of the application for constructing URLs.

Note that future versions of Yesod may add new constructors without bumping the major version number. As a result, you should not pattern match on Approot values.

Constructors

ApprootRelative
No application root.

ApprootStatic !Text

ApprootMaster !(master -> Text)

ApprootRequest !(master -> Request -> Text)

type ResolvedApproot = Text

data AuthResult

Constructors

Authorized

AuthenticationRequired

Unauthorized !Text

data ScriptLoadPosition master

Constructors

BottomOfBody

BottomOfHeadBlocking

BottomOfHeadAsync !(BottomOfHeadAsync master)

type BottomOfHeadAsync master

Arguments

= [Text]
urls to load asynchronously

-> Maybe (HtmlUrl (Route master))
widget of js to run on async completion

-> HtmlUrl (Route master)
widget to insert at the bottom of head

type Texts = [Text]

newtype WaiSubsite

Wrap up a normal WAI application as a Yesod subsite. Ignore parent site's middleware and isAuthorized.

Constructors

WaiSubsite

Fields

runWaiSubsite :: Application

newtype WaiSubsiteWithAuth

Like WaiSubsite, but applies parent site's middleware and isAuthorized.

Since: 1.4.34

Constructors

WaiSubsiteWithAuth

Fields

runWaiSubsiteWithAuth :: Application

data RunHandlerEnv child site

Constructors

RunHandlerEnv

Fields

rheRender :: !(Route site -> [(Text, Text)] -> Text)

rheRoute :: !(Maybe (Route child))

rheRouteToMaster :: !(Route child -> Route site)

rheSite :: !site

rheChild :: !child

rheUpload :: !(RequestBodyLength -> FileUpload)

rheLog :: !(Loc -> LogSource -> LogLevel -> LogStr -> IO ())

rheOnError :: !(ErrorResponse -> YesodApp)

How to respond when an error is thrown internally.

Since 1.2.0

rheMaxExpires :: !Text

data HandlerData child site

Constructors

HandlerData

Fields

handlerRequest :: !YesodRequest

handlerEnv :: !(RunHandlerEnv child site)

handlerState :: !(IORef GHState)

handlerResource :: !InternalState

data YesodRunnerEnv site

Constructors

YesodRunnerEnv

Fields

yreLogger :: !Logger

yreSite :: !site

yreSessionBackend :: !(Maybe SessionBackend)

yreGen :: !(IO Int)

Generate a random number

yreGetMaxExpires :: !(IO Text)

data YesodSubRunnerEnv sub parent

Constructors

YesodSubRunnerEnv

Fields

ysreParentRunner :: !(ParentRunner parent)

ysreGetSub :: !(parent -> sub)

ysreToParentRoute :: !(Route sub -> Route parent)

ysreParentEnv :: !(YesodRunnerEnv parent)

type ParentRunner parent = HandlerFor parent TypedContent -> YesodRunnerEnv parent -> Maybe (Route parent) -> Application

newtype HandlerFor site a

A generic handler monad, which can have a different subsite and master site. We define a newtype for better error message.

Constructors

HandlerFor

Fields

unHandlerFor :: HandlerData site site -> IO a

data GHState

Constructors

GHState

Fields

ghsSession :: !SessionMap

ghsRBC :: !(Maybe RequestBodyContents)

ghsIdent :: !Int

ghsCache :: !TypeMap

ghsCacheBy :: !KeyedTypeMap

ghsHeaders :: !(Endo [Header])

type YesodApp = YesodRequest -> ResourceT IO YesodResponse

An extension of the basic WAI Application datatype to provide extra features needed by Yesod. Users should never need to use this directly, as the HandlerT monad and template haskell code should hide it away.

newtype WidgetFor site a

A generic widget, allowing specification of both the subsite and master site datatypes. While this is simply a WriterT, we define a newtype for better error messages.

Constructors

WidgetFor

Fields

unWidgetFor :: WidgetData site -> IO a

data WidgetData site

Constructors

WidgetData

Fields

wdRef :: !(IORef (GWData (Route site)))

wdHandler :: !(HandlerData site site)

tellWidget :: GWData (Route site) -> WidgetFor site ()

type RY master = Route master -> [(Text, Text)] -> Text

newtype CssBuilder

Newtype wrapper allowing injection of arbitrary content into CSS.

Usage:

toWidget $ CssBuilder "p { color: red }"

Since: 1.1.3

Constructors

CssBuilder

Fields

unCssBuilder :: Builder

data PageContent url

Content for a web page. By providing this datatype, we can easily create generic site templates, which would have the type signature:

PageContent url -> HtmlUrl url

Constructors

PageContent

Fields

pageTitle :: !Html

pageHead :: !(HtmlUrl url)

pageBody :: !(HtmlUrl url)

data Content

Constructors

ContentBuilder !Builder !(Maybe Int)
The content and optional content length.

ContentSource !(ConduitT () (Flush Builder) (ResourceT IO) ())

ContentFile !FilePath !(Maybe FilePart)

ContentDontEvaluate !Content

data TypedContent

Constructors

TypedContent !ContentType !Content

type RepHtml = Html

Deprecated: Please use Html instead

newtype RepJson

Constructors

RepJson Content

newtype RepPlain

Constructors

RepPlain Content

newtype RepXml

Constructors

RepXml Content

type ContentType = ByteString

data JSONResponse a where

Wrapper around types so that Handlers can return a domain type, even when the data will eventually be encoded as JSON. Example usage in a type signature:

postSignupR :: Handler (JSONResponse CreateUserResponse)

And in the implementation:

return $ JSONResponse $ CreateUserResponse userId

Since: 1.6.14

Constructors

JSONResponse :: ToJSON a => a -> JSONResponse a

newtype DontFullyEvaluate a

Prevents a response body from being fully evaluated before sending the request.

Since 1.1.0

Constructors

DontFullyEvaluate

Fields

unDontFullyEvaluate :: a

data ErrorResponse

Responses to indicate some form of an error occurred.

Constructors

NotFound

InternalError !Text

InvalidArgs ![Text]

NotAuthenticated

PermissionDenied !Text

BadMethod !Method

data Header

Headers to be added to a Result.

Constructors

AddCookie !SetCookie

DeleteCookie !ByteString !ByteString
name and path

Header !(CI ByteString) !ByteString
key and value

data Location url

Constructors

Local !url

Remote !Text

newtype UniqueList x

A diff list that does not directly enforce uniqueness. When creating a widget Yesod will use nub to make it unique.

Constructors

UniqueList ([x] -> [x])

data Script url

Constructors

Script

Fields

scriptLocation :: !(Location url)

scriptAttributes :: ![(Text, Text)]

data Stylesheet url

Constructors

Stylesheet

Fields

styleLocation :: !(Location url)

styleAttributes :: ![(Text, Text)]

newtype Title

Constructors

Title

Fields

unTitle :: Html

newtype Head url

Constructors

Head (HtmlUrl url)

newtype Body url

Constructors

Body (HtmlUrl url)

type CssBuilderUrl a = (a -> [(Text, Text)] -> Text) -> Builder

data GWData a

Constructors

GWData

Fields

gwdBody :: !(Body a)

gwdTitle :: !(Last Title)

gwdScripts :: !(UniqueList (Script a))

gwdStylesheets :: !(UniqueList (Stylesheet a))

gwdCss :: !(Map (Maybe Text) (CssBuilderUrl a))

gwdJavascript :: !(Maybe (JavascriptUrl a))

gwdHead :: !(Head a)

data HandlerContents

Constructors

HCContent !Status !TypedContent

HCError !ErrorResponse

HCSendFile !ContentType !FilePath !(Maybe FilePart)

HCRedirect !Status !Text

HCCreated !Text

HCWai !Response

HCWaiApp !Application

data Logger

Constructors

Logger

Fields

loggerSet :: !LoggerSet

loggerDate :: !DateCacheGetter

loggerPutStr :: Logger -> LogStr -> IO ()

newtype SubHandlerFor sub master a

A handler monad for subsite

Since: 1.6.0

Constructors

SubHandlerFor

Fields

unSubHandlerFor :: HandlerData sub master -> IO a