Yesod.Core.Widget

Description

Widgets combine HTML with JS and CSS dependencies with a unique identifier generator, allowing you to create truly modular HTML components.

Datatype

data WidgetT site m a

A generic widget, allowing specification of both the subsite and master site datatypes. While this is simply a WriterT, we define a newtype for better error messages.

data PageContent url

Content for a web page. By providing this datatype, we can easily create generic site templates, which would have the type signature:

 PageContent url -> HtmlUrl url

Constructors

PageContent

Fields

pageTitle :: Html

pageHead :: HtmlUrl url

pageBody :: HtmlUrl url

Special Hamlet quasiquoter/TH for Widgets

whamlet :: QuasiQuoter

whamletFile :: FilePath -> Q Exp

ihamletToRepHtml :: (MonadHandler m, RenderMessage (HandlerSite m) message) => HtmlUrlI18n message (Route (HandlerSite m)) -> m Html

Deprecated: Please use ihamletToHtml instead

Wraps the Content generated by hamletToContent in a RepHtml.

ihamletToHtml :: (MonadHandler m, RenderMessage (HandlerSite m) message) => HtmlUrlI18n message (Route (HandlerSite m)) -> m Html

Wraps the Content generated by hamletToContent in a RepHtml.

Since 1.2.1

Convert to Widget

class ToWidget site a where

Methods

toWidget :: (MonadWidget m, HandlerSite m ~ site) => a -> m ()

class ToWidgetHead site a where

Methods

toWidgetHead :: (MonadWidget m, HandlerSite m ~ site) => a -> m ()

class ToWidgetBody site a where

Methods

toWidgetBody :: (MonadWidget m, HandlerSite m ~ site) => a -> m ()

class ToWidgetMedia site a where

Allows adding some CSS to the page with a specific media type.

Since 1.2

Methods

toWidgetMedia

Arguments

:: (MonadWidget m, HandlerSite m ~ site)

=> Text
media value

-> a

-> m ()

Add the given content to the page, but only for the given media type.

Since 1.2

Creating

Head of page

setTitle :: MonadWidget m => Html -> m ()

Set the page title. Calling setTitle multiple times overrides previously set values.

setTitleI :: (MonadWidget m, RenderMessage (HandlerSite m) msg) => msg -> m ()

Set the page title. Calling setTitle multiple times overrides previously set values.

CSS

addStylesheet :: MonadWidget m => Route (HandlerSite m) -> m ()

Link to the specified local stylesheet.

addStylesheetAttrs :: MonadWidget m => Route (HandlerSite m) -> [(Text, Text)] -> m ()

Link to the specified local stylesheet.

addStylesheetRemote :: MonadWidget m => Text -> m ()

Link to the specified remote stylesheet.

addStylesheetRemoteAttrs :: MonadWidget m => Text -> [(Text, Text)] -> m ()

Link to the specified remote stylesheet.

addStylesheetEither :: MonadWidget m => Either (Route (HandlerSite m)) Text -> m ()

newtype CssBuilder

Newtype wrapper allowing injection of arbitrary content into CSS.

Usage:

 toWidget $ CssBuilder "p { color: red }"

Since: 1.1.3

Constructors

CssBuilder

Fields

unCssBuilder :: Builder

Javascript

addScript :: MonadWidget m => Route (HandlerSite m) -> m ()

Link to the specified local script.

addScriptAttrs :: MonadWidget m => Route (HandlerSite m) -> [(Text, Text)] -> m ()

Link to the specified local script.

addScriptRemote :: MonadWidget m => Text -> m ()

Link to the specified remote script.

addScriptRemoteAttrs :: MonadWidget m => Text -> [(Text, Text)] -> m ()

Link to the specified remote script.

addScriptEither :: MonadWidget m => Either (Route (HandlerSite m)) Text -> m ()

Subsites

widgetToParentWidget :: MonadIO m => WidgetT child IO a -> HandlerT child (HandlerT parent m) (WidgetT parent m a)

handlerToWidget :: Monad m => HandlerT site m a -> WidgetT site m a

Internal

whamletFileWithSettings :: HamletSettings -> FilePath -> Q Exp

asWidgetT :: WidgetT site m () -> WidgetT site m ()